fix(#26046): cache detecting logic moved to Image component.
