+++ src/components/Image/index.js
@@ -1,4 +1,4 @@
-import React, {useEffect, useMemo} from 'react';
+import React, {useEffect, useMemo, useRef, useState} from 'react';
 import {Image as RNImage} from 'react-native';
 import {withOnyx} from 'react-native-onyx';
 import lodashGet from 'lodash/get';
@@ -8,7 +8,10 @@ import {defaultProps, imagePropTypes} from './imagePropTypes';
 import RESIZE_MODES from './resizeModes';
 
 function Image(props) {
-    const {source: propsSource, isAuthTokenRequired, onLoad, session} = props;
+    const {source: propsSource, isAuthTokenRequired, onLoad, session, onLoadStart = () => {}, onLoadEnd = () => {}} = props;
+
+    const [isLoading, setIsLoading] = useState(false);
+    const isLoadedRef = useRef(null);
     /**
      * Check if the image source is a URL - if so the `encryptedAuthToken` is appended
      * to the source.
@@ -41,14 +44,34 @@ function Image(props) {
         });
     }, [onLoad, source]);
 
+    /** Delay the loader to detect whether the image is being loaded from the cache or the internet. */
+    useEffect(() => {
+        if (isLoadedRef.current || !isLoading) {
+            return;
+        }
+        const timeout = _.delay(() => {
+            if (!isLoading || isLoadedRef.current) {
+                return;
+            }
+            onLoadStart();
+        }, 200);
+        return () => clearTimeout(timeout);
+    }, [isLoading, onLoadStart]);
+
     // Omit the props which the underlying RNImage won't use
-    const forwardedProps = _.omit(props, ['source', 'onLoad', 'session', 'isAuthTokenRequired']);
+    const forwardedProps = _.omit(props, ['source', 'onLoad', 'session', 'isAuthTokenRequired', 'onLoadStart', 'onLoadEnd']);
 
     return (
         <RNImage
             // eslint-disable-next-line react/jsx-props-no-spreading
             {...forwardedProps}
             source={source}
+            onLoadStart={() => setIsLoading(true)}
+            onLoadEnd={() => {
+                onLoadEnd();
+                setIsLoading(false);
+                isLoadedRef.current = true;
+            }}
         />
     );
 }
